+++ libs/core/langchain_core/globals/__init__.py
@@ -19,7 +19,7 @@ _llm_cache: Optional["BaseCache"] = None
 def set_verbose(value: bool) -> None:
     """Set a new value for the `verbose` global setting."""
     try:
-        import langchain
+        import langchain  # type: ignore[import]
 
         # We're about to run some deprecated code, don't report warnings from it.
         # The user called the correct (non-deprecated) code path and shouldn't get warnings.
@@ -47,7 +47,7 @@ def set_verbose(value: bool) -> None:
 def get_verbose() -> bool:
     """Get the value of the `verbose` global setting."""
     try:
-        import langchain
+        import langchain  # type: ignore[import]
 
         # We're about to run some deprecated code, don't report warnings from it.
         # The user called the correct (non-deprecated) code path and shouldn't get warnings.
@@ -80,7 +80,7 @@ def get_verbose() -> bool:
 def set_debug(value: bool) -> None:
     """Set a new value for the `debug` global setting."""
     try:
-        import langchain
+        import langchain  # type: ignore[import]
 
         # We're about to run some deprecated code, don't report warnings from it.
         # The user called the correct (non-deprecated) code path and shouldn't get warnings.
@@ -106,7 +106,7 @@ def set_debug(value: bool) -> None:
 def get_debug() -> bool:
     """Get the value of the `debug` global setting."""
     try:
-        import langchain
+        import langchain  # type: ignore[import]
 
         # We're about to run some deprecated code, don't report warnings from it.
         # The user called the correct (non-deprecated) code path and shouldn't get warnings.
@@ -137,7 +137,7 @@ def get_debug() -> bool:
 def set_llm_cache(value: Optional["BaseCache"]) -> None:
     """Set a new LLM cache, overwriting the previous value, if any."""
     try:
-        import langchain
+        import langchain  # type: ignore[import]
 
         # We're about to run some deprecated code, don't report warnings from it.
         # The user called the correct (non-deprecated) code path and shouldn't get warnings.
@@ -165,7 +165,7 @@ def set_llm_cache(value: Optional["BaseCache"]) -> None:
 def get_llm_cache() -> "BaseCache":
     """Get the value of the `llm_cache` global setting."""
     try:
-        import langchain
+        import langchain  # type: ignore[import]
 
         # We're about to run some deprecated code, don't report warnings from it.
         # The user called the correct (non-deprecated) code path and shouldn't get warnings.
