+++ libs/core/langchain_core/language_models/__init__.py
@@ -1,4 +1,8 @@
-from langchain_core.language_models.base import BaseLanguageModel, LanguageModelInput
+from langchain_core.language_models.base import (
+    BaseLanguageModel,
+    LanguageModelInput,
+    get_tokenizer,
+)
 from langchain_core.language_models.chat_models import BaseChatModel, SimpleChatModel
 from langchain_core.language_models.llms import LLM, BaseLLM
 
@@ -9,4 +13,5 @@ __all__ = [
     "BaseLLM",
     "LLM",
     "LanguageModelInput",
+    "get_tokenizer",
 ]
