+++ libs/core/langchain_core/language_models/chat_models.py
@@ -39,7 +39,7 @@ from langchain_core.outputs import (
     LLMResult,
     RunInfo,
 )
-from langchain_core.prompts import ChatPromptValue, PromptValue, StringPromptValue
+from langchain_core.prompt_values import ChatPromptValue, PromptValue, StringPromptValue
 from langchain_core.pydantic_v1 import Field, root_validator
 from langchain_core.runnables import RunnableConfig
