+++ libs/core/langchain_core/language_models/llms.py
@@ -50,7 +50,7 @@ from langchain_core.language_models.base import BaseLanguageModel, LanguageModel
 from langchain_core.load import dumpd
 from langchain_core.messages import AIMessage, BaseMessage, get_buffer_string
 from langchain_core.outputs import Generation, GenerationChunk, LLMResult, RunInfo
-from langchain_core.prompts import ChatPromptValue, PromptValue, StringPromptValue
+from langchain_core.prompt_values import ChatPromptValue, PromptValue, StringPromptValue
 from langchain_core.pydantic_v1 import Field, root_validator, validator
 from langchain_core.runnables import RunnableConfig, get_config_list
