+++ libs/core/langchain_core/prompts/pipeline.py
@@ -1,8 +1,8 @@
 from typing import Any, Dict, List, Tuple
 
+from langchain_core.prompt_values import PromptValue
 from langchain_core.prompts.base import BasePromptTemplate
 from langchain_core.prompts.chat import BaseChatPromptTemplate
-from langchain_core.prompts.value import PromptValue
 from langchain_core.pydantic_v1 import root_validator
