+++ libs/core/langchain_core/prompts/string.py
@@ -4,11 +4,10 @@ from __future__ import annotations
 import warnings
 from abc import ABC
 from string import Formatter
-from typing import Any, Callable, Dict, List, Literal, Set
+from typing import Any, Callable, Dict, List, Set
 
-from langchain_core.messages import BaseMessage, HumanMessage
+from langchain_core.prompt_values import PromptValue, StringPromptValue
 from langchain_core.prompts.base import BasePromptTemplate
-from langchain_core.prompts.value import PromptValue
 from langchain_core.utils.formatting import formatter
 
 
@@ -147,22 +146,6 @@ def get_template_variables(template: str, template_format: str) -> List[str]:
         raise ValueError(f"Unsupported template format: {template_format}")
 
     return sorted(input_variables)
-
-
-class StringPromptValue(PromptValue):
-    """String prompt value."""
-
-    text: str
-    """Prompt text."""
-    type: Literal["StringPromptValue"] = "StringPromptValue"
-
-    def to_string(self) -> str:
-        """Return prompt as string."""
-        return self.text
-
-    def to_messages(self) -> List[BaseMessage]:
-        """Return prompt as messages."""
-        return [HumanMessage(content=self.text)]
 
 
 class StringPromptTemplate(BasePromptTemplate, ABC):
