+++ libs/core/langchain_core/tracers/base.py
@@ -9,7 +9,7 @@ from uuid import UUID
 
 from tenacity import RetryCallState
 
-from langchain_core.callbacks import BaseCallbackHandler
+from langchain_core.callbacks.base import BaseCallbackHandler
 from langchain_core.documents import Document
 from langchain_core.exceptions import TracerException
 from langchain_core.load import dumpd
