+++ libs/core/tests/unit_tests/prompts/test_chat.py
@@ -10,6 +10,7 @@ from langchain_core.messages import (
     SystemMessage,
     get_buffer_string,
 )
+from langchain_core.prompt_values import ChatPromptValue
 from langchain_core.prompts import PromptTemplate
 from langchain_core.prompts.chat import (
     AIMessagePromptTemplate,
@@ -17,7 +18,6 @@ from langchain_core.prompts.chat import (
     ChatMessage,
     ChatMessagePromptTemplate,
     ChatPromptTemplate,
-    ChatPromptValue,
     HumanMessagePromptTemplate,
     SystemMessagePromptTemplate,
     _convert_to_message,
