+++ libs/langchain/langchain/llms/anthropic.py
@@ -13,7 +13,7 @@ from typing import (
 
 from langchain_core.language_models import BaseLanguageModel
 from langchain_core.outputs import GenerationChunk
-from langchain_core.prompts import PromptValue
+from langchain_core.prompt_values import PromptValue
 from langchain_core.pydantic_v1 import Field, SecretStr, root_validator
 from langchain_core.utils import (
     check_package_version,
