+++ libs/langchain/langchain/prompts/base.py
@@ -1,7 +1,6 @@
-from langchain_core.prompts import BasePromptTemplate
-from langchain_core.prompts.base import (
+from langchain_core.prompts import (
+    BasePromptTemplate,
     StringPromptTemplate,
-    StringPromptValue,
     check_valid_template,
     get_template_variables,
     jinja2_formatter,
@@ -13,7 +12,6 @@ __all__ = [
     "validate_jinja2",
     "check_valid_template",
     "get_template_variables",
-    "StringPromptValue",
     "StringPromptTemplate",
     "BasePromptTemplate",
 ]
